Extract ephemeraldb only tests
